Creating a new patient continued
Still have the same name being assigned to many of the patients

+++ FINAL_PROJECT/CS273_FINAL_PROJECT_HospitalSimulator/CS273_FINAL_PROJECT_HospitalSimulator/RandomsGenerator.h
@@ -15,7 +15,7 @@ struct Random {
 	}
 
 	//return random double for nextInt to utilize
-	int nextDouble() {
+	double nextDouble() {
 		return double(rand()) / RAND_MAX;
 	}
 
@@ -34,7 +34,6 @@ struct Random {
 			priority = rand() % 10 + 1; 
 			return priority;}
 	}
-
 
 
 };
